refactor(app): move analytics out of the homepage feature
app/_home/ is the directory for things only the homepage uses, but
analytics.ts sat there while app/_components/PrivacySafeAnalytics.tsx
imported it across the boundary -- so the root layout, which has
nothing to do with searching for a commit, reached into the homepage
feature on every page load. It now sits in app/_lib/ with the other
shared modules, which leaves app/_home/ meaning what its name claims.

The import stays relative rather than aliased: every non-route file in
this app reaches _lib relatively, and only route files under app/api/
use @/app/_lib. Making this one the exception would have traded a
wrong location for an inconsistent import.

Also records the rule that decides this for the next module, and
catches the layout list up on app/api/csp-report/route.ts and
app/_lib/e2eCommitSearchFixtures.ts, plus four comments still pointing
at pre-_lib paths.

Entire-Checkpoint: 01M08H30WJYAB1DP81T73A74VP
