+++ app/_home/searchExecution.ts
@@ -2,7 +2,7 @@ import type { Dispatch, SetStateAction, TransitionStartFunction } from "react";
 import { getCommits } from "../_lib/actions";
 import type { CommitData } from "../_lib/commitTypes";
 import { getUsernameValidationMessage, normalizeGitHubUsername } from "../_lib/username";
-import { trackAppEvent } from "./analytics";
+import { trackAppEvent } from "../_lib/analytics";
 import { MAX_RECENT_SEARCHES } from "./constants";
 import { saveStoredRecentSearches } from "./recentSearches";
 import { updateSharedSearchUrl } from "./sharedSearch";
