+++ app/_lib/analytics.ts
